test: soumission du profil avec données correcte ok

+++ tests/Controller/ProfilControllerTest.php
@@ -9,12 +9,13 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\DomCrawler\Crawler;
 use App\Entity\User;
 use App\Repository\UserRepository;
+use Hautelook\AliceBundle\PhpUnit\ReloadDatabaseTrait;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Vich\UploaderBundle\Templating\Helper\UploaderHelper;
 
 class ProfilControllerTest extends WebTestCase
 {
-    use RefreshDatabaseTrait;
+    use ReloadDatabaseTrait;
     use LoadFixtureTrait;
 
     private ?KernelBrowser $client;
@@ -23,6 +24,7 @@ class ProfilControllerTest extends WebTestCase
     protected function setUp(): void
     {
         $this->client = $this->createClient();
+        $this->pathMockFile = [];
     }
 
     public function testPageProfilNotAccessIfUserIsAnonymous(): void
@@ -50,22 +52,19 @@ class ProfilControllerTest extends WebTestCase
      */
     public function testSubmitFormSuccess(array $formData): void
     {
-        $path = null;
-        if (array_key_exists('file_info', $formData['profil'])) {
-            $path = $this->mockFile($formData['profil']['file_info']['mimetype'], $formData['profil']['file_info']['filename']);
-            $filename = $formData['profil']['file_info']['filename'];
-            $formData['profil']['file']['file'] = new UploadedFile($path, $filename);
-            unset($formData['profil']['file_info']);
-            $this->pathMockFile[] = $path;
-        }
+
+        $formData = $this->handleFormData($formData);
 
         $authenticatedUser = $this->getFixtures()['user_credentials_ok'];
         $this->client->loginUser($authenticatedUser);
         /** @var Crawler */
         $crawler = $this->client->request('GET', '/profil');
+        $this->assertResponseIsSuccessful();
 
         $form = $crawler->selectButton('Modifier')->form($formData);
         $this->client->submit($form);
+
+
         /** @var UserRepository */
         $userRepository = $this->getContainer()->get(UserRepository::class);
         /** @var User */
@@ -74,16 +73,33 @@ class ProfilControllerTest extends WebTestCase
         $this->assertInstanceOf(User::class, $userExpected);
         $this->assertEquals(strtoupper($formData['profil']['nom']), $userExpected->getNom());
         $this->assertEquals(ucwords($formData['profil']['prenom']), $userExpected->getPrenom());
-        /** @var UploaderHelper */
-        $uploaderHelper = $this->getContainer()->get(UploaderHelper::class);
-        $pathFileUploaded = $uploaderHelper->asset($userExpected, 'file');
-        $this->pathMockFile[] = $pathFileUploaded;
-        // $filenameInBdd = $userExpected->getImageName();
-        $this->assertFileExists($path);
+
+        if ($userExpected->getImageName()) {
+            $dirFileUploaded = $this->getContainer()->getParameter('path_uploaded_image_users');
+            $pathFileUploaded = $dirFileUploaded . DIRECTORY_SEPARATOR . $userExpected->getImageName();
+            $this->pathMockFile[] = $pathFileUploaded;
+            $this->assertFileExists($pathFileUploaded);
+        }
+
         $this->assertResponseStatusCodeSame(302);
         $this->client->followRedirect();
-
+        // $this->assertResponseRedirects('/profil');
         // $this->assertSelectorExists('.alert-success');
+    }
+    /**
+     * modification de donnée et suppression du file_info
+     */
+    private function handleFormData(array $formData): array
+    {
+        $path = null;
+        if (array_key_exists('file_info', $formData['profil'])) {
+            $path = $this->mockFile($formData['profil']['file_info']['mimetype'], $formData['profil']['file_info']['filename']);
+            $filename = $formData['profil']['file_info']['filename'];
+            $formData['profil']['file']['file'] = new UploadedFile($path, $filename);
+            unset($formData['profil']['file_info']);
+            $this->pathMockFile[] = $path;
+        }
+        return $formData;
     }
 
     private function mockFile(string $mimeType, string $filename): string
@@ -151,9 +167,10 @@ class ProfilControllerTest extends WebTestCase
     protected function tearDown(): void
     {
         parent::tearDown();
-
-        foreach ($this->pathMockFile as $path) {
-            unlink($path);
+        if ($this->pathMockFile) {
+            foreach ($this->pathMockFile as $path) {
+                unlink($path);
+            }
         }
         $this->pathMockFile = null;
         $this->client = null;
